Added arrow styles into diagram schema.
The DiagramSchema class now stores the ArrowStyle[] used by the
temporal components, instead of the temporal controls panel. It also
maintains this information in the preferences, which means these styles
are now part of the session management.

The XMLization of the arrows has been changed to refer to the styles
by number instead of storing them in the CSX. The latter approach didn't
really work anyway, since these styles couldn't be changed properly
after a reload.


git-svn-id: svn://svn.code.sf.net/p/toscanaj/code/trunk@1637 267469b6-68fb-4578-8a4b-5844dc4013a7