Objects now remember their position in the context.
The FCAObjectImplementation class is able to store a context position now
and implement Comparable using this order.

The context editor does store this (through a method on 
ContextImplementation), the relevant other parts of the program (e.g.
the diagram to context converter) check if the objects are Comparable
and inserts them in order if they are.

Attributes don't behave like this yet, they will be changed to use the same
class.


git-svn-id: svn://svn.code.sf.net/p/toscanaj/code/trunk@1643 267469b6-68fb-4578-8a4b-5844dc4013a7