Renamed FCAObject into FCAElement (and derived classes).
The interfaces/classes FCAObject, WritableFCAObject and 
FCAObjectImplementation have been renames to use "Element" instead
of "Object" in their names (s/Object/Element/).

This is done as preparation of sharing the interfaces and classes with the
attributes. Once this is done, "object" will be nothing more but a role in
the context, not anything particular for that class anymore. Thus the
more neutral "Element" -- "Object" has too strong connotations in FCA.


git-svn-id: svn://svn.code.sf.net/p/toscanaj/code/trunk@1644 267469b6-68fb-4578-8a4b-5844dc4013a7