Siena now calculated a many-valued context in case there is none.
After loading a CSX file it is checked if a many-valued context was found.
If not, a new one will created based on all the single-valued contexts
from the diagrams.

One thing that occured to me is that drawing performance is very bad.
We need to check the context editor for performance leaks...


git-svn-id: svn://svn.code.sf.net/p/toscanaj/code/trunk@1655 267469b6-68fb-4578-8a4b-5844dc4013a7