Added some work in progress for datatypes.
Packages org.tockit.datatypes and org.tockit.datatypes.xsd.

This is not yet finished and might not be used in the next version, but I
do not want to lose it or create an extra branch for it. The new files are
in the two packages mentioned above, there are no other changes.


git-svn-id: svn://svn.code.sf.net/p/toscanaj/code/trunk@1669 267469b6-68fb-4578-8a4b-5844dc4013a7