Bugfix: exceptions when using SQL queries resolved.
When the infimum of the realized lattice has an empty contingent the
queries are executed (since the concept is realized), but the SQL will
return NULL values. This caused the labels to throw exceptions, the new
version will catch this case and return NULL for the list instead of NULLs
in the list, which means the label does not get displayed.


git-svn-id: svn://svn.code.sf.net/p/toscanaj/code/trunk@1684 267469b6-68fb-4578-8a4b-5844dc4013a7