Lattice creation now uses bitsets.
Finally did a task postponed quite often and refactored the lattice creation algorithm
to use BitSets instead of ListSets. Order should be still ensured, but it is a lot faster
now.

The one benchmark I did (importing a Cernato file 10 times) went down from an
average ~5600 millis to an average of ~400 millis, i.e. a speedup of 14x. That
of course includes IO and diagram layout, so the actual speedup of the algorithm
is likely to be quite a bit higher.


git-svn-id: svn://svn.code.sf.net/p/toscanaj/code/trunk@1736 267469b6-68fb-4578-8a4b-5844dc4013a7