A decent table view for Siena.
I reimplemented the TableView and related classes from scratch, this time based on
JTable. This means we are adjusted to the look and feel, we do have column width
resizing and column reordering and the cell editing framework from JTable.

Adding the RowHeader is a bit tricky in Swing, but I found a good description on
this site: http://www.chka.de/swing/table/row-headers/JTable.html

A lot of the new code is based on this (up to copy&paste, the assumption is that
the code on the site is public domain).


git-svn-id: svn://svn.code.sf.net/p/toscanaj/code/trunk@1759 267469b6-68fb-4578-8a4b-5844dc4013a7