Automated classpath generation for JARs, added JARs for Elba and Siena.
The Ant build file now creates two small JARs for Elba and Siena, containing
nothing but a manifest telling Java how to start the programs. This way we do not
need to provide an external classpath anymore. And people can start all applications
through jars.

Additionally the classpath in the manifests is now created by Ant, which means we
never need to edit any classpath anymore.


git-svn-id: svn://svn.code.sf.net/p/toscanaj/code/trunk@1787 267469b6-68fb-4578-8a4b-5844dc4013a7