Bugfix: confusion with outer nodes of nested nodes.
There is a name conflict in the code of nested diagram nodes: it has an
outer node in the sense that it represents one node of the outer diagram
in the nesting. But it also has an outer node in the sense that the outer
diagram could be nested in something else. This was confused in one place,
causing the nesting list to be wrong.
itself


git-svn-id: svn://svn.code.sf.net/p/toscanaj/code/trunk@1837 267469b6-68fb-4578-8a4b-5844dc4013a7