Added nested diagrams to Siena.
Using the context menu of the diagram list, the user can
now select a second diagram to be nested into the
diagram currently displayed.

Filter and Ideal movement are still a bit buggy (they don't
work on the inner diagram), and the temporal views
operate on the outer diagram instead on the nested
one, otherwise it seems ok.


git-svn-id: svn://svn.code.sf.net/p/toscanaj/code/trunk@1846 267469b6-68fb-4578-8a4b-5844dc4013a7