Added ability to change end points of transition arrows.
If the user drags in the first or last quarter of a transition arrow, only the corresponding tip will be moved. This way the arrow length and angle can be changed if required.

Dragging in the central part still shows the old behaviour of moving the arrow as a whole.


git-svn-id: svn://svn.code.sf.net/p/toscanaj/code/trunk@1868 267469b6-68fb-4578-8a4b-5844dc4013a7