Fixed regression: connections to the embedded DB work again.
HSQLDB changed behaviour with version 1.7.2 and above: the database would not shut down after the last connection was dropped, thus keeping old tables when we expected them to be removed.

With 1.8.0 they introduced a connection property to enforce the old behaviour, which we now use. This also means upgrading to 1.8.0.


git-svn-id: svn://svn.code.sf.net/p/toscanaj/code/trunk@1871 267469b6-68fb-4578-8a4b-5844dc4013a7