Added feature to DataDump class to create images for all diagrams.
A new CLI option specifying the file type can be used to let DataDump create images for each diagram. This is still a bit of a hack, e.g. the image size is still a hardcoded 1280x1024 pixels and it is definitely not the most performant since the DB is queried twice -- once for the XML export, once for the images.


git-svn-id: svn://svn.code.sf.net/p/toscanaj/code/trunk@1872 267469b6-68fb-4578-8a4b-5844dc4013a7