Bugfix: HTML database viewer now does the right queries.
The single fields were queried along with the aggregate fields, which was superflous at best. Hsqldb in its latest incarnation didn't like it, which seems to have broken the default view of the PCTest example.

The aggregates are now no longer combined with the single fields, so this should be ok.

git-svn-id: svn://svn.code.sf.net/p/toscanaj/code/trunk@1896 267469b6-68fb-4578-8a4b-5844dc4013a7