The table now always contains an additional row and column to add new objects/attributes.
This is the same approach as taken in Cernato: to avoid not having any table to see, an additional row and an additional column are added, which can be used to add new objects/attributes. The latter functionality is not implemented yet, but the labels are already there.

While doing this the focus/selection highlighting in the row header was removed.

git-svn-id: svn://svn.code.sf.net/p/toscanaj/code/trunk@1909 267469b6-68fb-4578-8a4b-5844dc4013a7