CSC import now supports LINE_DIAGRAM elements.
The original code more than 4 years ago imported only LINE_DIAGRAM elements, then it changed to only CONCRETE_SCALE elements, but that fails for some examples provided by users. The new version imports both types, which should hopefully work for most use cases -- due to a lack of examples as well as the specification it's a bit of guesswork.

The CSCParser in ToscanaJ is also renamed to CSCImport and does not extend org.tockit.conscript.parser.CSCParser anymore, since it is not really a parser by itself.

git-svn-id: svn://svn.code.sf.net/p/toscanaj/code/trunk@1969 267469b6-68fb-4578-8a4b-5844dc4013a7