OAL import now supports adding attributes without objects.
Whenever an OAL file has a line starting with the colon, the following attributes are just added to the context without creating an object. This fixes a bug where such a line would cause exceptions in the display code and it also allows adding attributes to a top concept without extent.

git-svn-id: svn://svn.code.sf.net/p/toscanaj/code/trunk@1977 267469b6-68fb-4578-8a4b-5844dc4013a7